+++ crates/ruma-client-api/CHANGELOG.md
@@ -6,7 +6,7 @@ Breaking changes:
   * Use `sync::sync_events::DeviceLists` instead
 * `fully_read` field in `read_marker::set_read_marker` is no longer required
   * Remove the `fully_read` argument from `read_marker::set_read_marker::Request::new`
-* Move `message::get_message_events::v3::Direction` to the root of the crate
+* Move `message::get_message_events::v3::Direction` to `ruma-common::api`
 * Make `push::set_pusher::v3::Request` use an enum to differentiate when deleting a pusher
   * Move `push::get_pushers::v3::Pusher` to `push` and make it use the new `PusherIds` type
   * Remove `push::set_pusher::v3::Pusher` and use the common type instead